Add Kiniesis consumer lambda

+++ Makefile
@@ -1,6 +1,6 @@
 LOCALSTACK_TMPDIR=./localstack-temp
 LOCALSTACK_ENDPOINT=http://localhost:4566 
-AWS=aws --endpoint ${LOCALSTACK_ENDPOINT}
+AWS=aws --endpoint ${LOCALSTACK_ENDPOINT} --cli-read-timeout 0 --cli-connect-timeout 0
 BUCKET_NAME=webpages
 BUS_NAME=webpages-bus
 RULE_NAME=webpages-event-rule
@@ -20,8 +20,11 @@ start-localstack: create-aws-profile
 create-bucket:
 	${AWS} s3api create-bucket --bucket ${BUCKET_NAME}
 
-create-lambda:
-	make -C scraper create-lambda
+create-scraper-lambda:
+	make -C scraper all
+
+create-eps-lambda:
+	make -C eps all
 
 create-eventbridge:
 	make -C eventbridge-rules all
@@ -30,14 +33,14 @@ create-kinesis-stream:
 	make -C kinesis/stream all
 
 stop:
-	$(eval CONTAINER_ID := $(shell docker ps | grep localstack | cut -d " " -f 1))
-	docker stop $(CONTAINER_ID)
-	docker rm   $(CONTAINER_ID)
+	$(eval CONTAINER_ID := $(shell docker ps | grep localstack | cut -d " " -f 1 ))
+	[[ ! -z "$(CONTAINER_ID)" ]] && docker stop $(CONTAINER_ID) || true
+	[[ ! -z "$(CONTAINER_ID)" ]] && docker rm   $(CONTAINER_ID) || true
 
 run: \
-	stop \
 	start-localstack \
 	create-bucket \
-	create-lambda \
+	create-scraper-lambda \
 	create-eventbridge \
-	create-kinesis-stream
+	create-kinesis-stream \
+	create-eps-lambda
